+++ src/gui/renderer.h
@@ -5,66 +5,65 @@
 #include <stack>
 #include <vector>
 
-#include "../common.h"
-#include "../sdl2.h"
+#include "sdl_headers.h"
 
 class Renderer {
 public:
 	Renderer(SDL_Renderer* ren);
 	~Renderer();
 
-	void line(i32 x1, i32 y1, i32 x2, i32 y2, u8 r, u8 g, u8 b, u8 a = 255, u8 width = 1);
+	void line(int x1, int y1, int x2, int y2, int r, int g, int b, int a = 255, int width = 1);
 	void curve(
-		i32 x1, i32 y1,
-		i32 x2, i32 y2,
-		i32 x3, i32 y3,
-		i32 x4, i32 y4,
-		u8 r, u8 g, u8 b, u8 a = 255
+		int x1, int y1,
+		int x2, int y2,
+		int x3, int y3,
+		int x4, int y4,
+		int r, int g, int b, int a = 255
 	);
 	void multiCurve(
-		const std::vector<i32>& positions,
-		u8 r, u8 g, u8 b, u8 a = 255
+		const std::vector<int>& positions,
+		int r, int g, int b, int a = 255
 	);
 	void roundRect(
-		i32 x, i32 y, i32 w, i32 h, i32 rad, u8 r, u8 g, u8 b, u8 a = 255
+		int x, int y, int w, int h, int rad, int r, int g, int b, int a = 255
 	);
-	void arc(i32 x, i32 y, i32 rad, i32 start, i32 end, u8 r, u8 g, u8 b, u8 a = 255);
+	void arc(int x, int y, int rad, int start, int end, int r, int g, int b, int a = 255);
 
-	void rect(i32 x, i32 y, i32 w, i32 h, u8 r, u8 g, u8 b, u8 a = 255, bool fill = false);
-	void triangle(i32 x1, i32 y1, i32 x2, i32 y2, i32 x3, i32 y3, u8 r, u8 g, u8 b, u8 a = 255);
-	void arrow(i32 x, i32 y, i32 w, i32 h, f32 angle, u8 r, u8 g, u8 b, u8 a = 255);
+	void rect(int x, int y, int w, int h, int r, int g, int b, int a = 255, bool fill = false);
+	void triangle(int x1, int y1, int x2, int y2, int x3, int y3, int r, int g, int b, int a = 255);
+	void arrow(int x, int y, int w, int h, float angle, int r, int g, int b, int a = 255);
 
-	void panel(i32 x, i32 y, i32 w, i32 h, i32 sx = 0, i32 sy = 0, f32 shadow = 0.0f);
-	void button(i32 x, i32 y, i32 w, i32 h, u32 state = 0);
-	void check(i32 x, i32 y, bool state);
-	void flatPanel(i32 x, i32 y, i32 w, i32 h, i32 sx = 0, i32 sy = 0, f32 shadow = 0.0f);
+	void panel(int x, int y, int w, int h, int sx = 0, int sy = 0, float shadow = 0.0f);
+	void button(int x, int y, int w, int h, int state = 0);
+	void check(int x, int y, bool state);
+	void flatPanel(int x, int y, int w, int h, int sx = 0, int sy = 0, float shadow = 0.0f);
 
-	void pushClipping(i32 x, i32 y, i32 w, i32 h);
+	void pushClipping(int x, int y, int w, int h);
 	void popClipping();
 
-	void text(i32 x, i32 y, const std::string& str, u8 r, u8 g, u8 b, u8 a = 255);
-	void textSmall(i32 x, i32 y, const std::string& str, u8 r, u8 g, u8 b, u8 a = 255);
-	u32 textWidth(const std::string& str) const { return str.size() * 8; }
+	void text(int x, int y, const std::string& str, int r, int g, int b, int a = 255);
+	void textSmall(int x, int y, const std::string& str, int r, int g, int b, int a = 255);
+	int textWidth(const std::string& str) const { return str.size() * 8; }
 
-	void patch(i32 x, i32 y, i32 w, i32 h, i32 sx, i32 sy, i32 sw, i32 sh, u32 pad = 5, i32 tpad = -1);
+	void patch(int x, int y, int w, int h, int sx, int sy, int sw, int sh, int pad = 5, int tpad = -1);
 
 	SDL_Renderer* sdlRenderer() { return m_ren; }
 
 private:
 	SDL_Renderer* m_ren;
 	SDL_Texture *m_skin, *m_font, *m_fontSmall;
-	u32 m_fontWidth, m_fontHeight;
-	u32 m_fontSmallWidth, m_fontSmallHeight;
+	int m_fontWidth, m_fontHeight;
+	int m_fontSmallWidth, m_fontSmallHeight;
 
 	std::stack<SDL_Rect> m_clipRects;
 
 	void skin(
-		i32 x, i32 y, i32 w, i32 h,
-		i32 sx, i32 sy, i32 sw, i32 sh
+		int x, int y, int w, int h,
+		int sx, int sy, int sw, int sh
 	);
 
-	void textGen(SDL_Texture* font, i32 fw, i32 fh, i32 x, i32 y, const std::string& str, u8 r, u8 g, u8 b, u8 a);
-	void putChar(SDL_Texture* font, i32 fw, i32 fh, i32 x, i32 y, char c, u8 r, u8 g, u8 b, u8 a = 0xFF);
+	void textGen(SDL_Texture* font, int fw, int fh, int x, int y, const std::string& str, int r, int g, int b, int a);
+	void putChar(SDL_Texture* font, int fw, int fh, int x, int y, char c, int r, int g, int b, int a = 0xFF);
 };
 
 #endif // SYG_RENDERER_H
