+++ src/node_canvas.h
@@ -1,17 +1,17 @@
 #ifndef SYG_NODE_CANVAS_H
 #define SYG_NODE_CANVAS_H
 
-#include "widget.h"
-#include "../../engine/node_logic.h"
+#include "gui/widgets/widget.h"
+#include "engine/node_logic.h"
 
 struct GNode {
-	i32 x{ 0 }, y{ 0 }, height{ 16 };
-	u32 node{ 0 };
+	int x{ 0 }, y{ 0 }, height{ 16 };
+	int node{ 0 };
 	bool selected{ false };
 };
 
 struct Link {
-	u32 src, dest, param;
+	int src, dest, param;
 	bool active{ false };
 };
 
@@ -20,8 +20,8 @@ public:
 	NodeCanvas();
 
 	template <typename T>
-	u32 create() {
-		u32 id = m_system->create<T>();
+	int create() {
+		int id = m_system->create<T>();
 		m_gnodes[id].x = 20;
 		m_gnodes[id].y = 20;
 		m_gnodes[id].node = id;
@@ -31,15 +31,15 @@ public:
 
 	virtual void onDraw(Renderer& renderer) override;
 
-	virtual void onClick(u8 button, i32 x, i32 y) override;
-	virtual void onMove(i32 x, i32 y) override;
-	virtual void onPress(u8 button, i32 x, i32 y) override;
-	virtual void onRelease(u8 button, i32 x, i32 y) override;
-	virtual void onKeyPress(u32 key, u32 mod) override;
-	virtual void onKeyRelease(u32 key, u32 mod) override;
+	virtual void onClick(int button, int x, int y) override;
+	virtual void onMove(int x, int y) override;
+	virtual void onPress(int button, int x, int y) override;
+	virtual void onRelease(int button, int x, int y) override;
+	virtual void onKeyPress(int key, int mod) override;
+	virtual void onKeyRelease(int key, int mod) override;
 
 	NodeSystem* system() { return m_system.get(); }
-	std::vector<u32> selected() { return m_selected; }
+	std::vector<int> selected() { return m_selected; }
 
 	void onSelect(const std::function<void(Node*)>& cb) { m_onSelect = cb; }
 	void onChange(const std::function<void()>& cb) { m_onChange = cb; }
@@ -69,13 +69,13 @@ private:
 		m_multiSelect{ false };
 	State m_state{ None };
 
-	std::vector<u32> m_selected;
+	std::vector<int> m_selected;
 
-	i32 m_px{ 0 }, m_py{ 0 }, m_sx{ 0 }, m_sy{ 0 },
+	int m_px{ 0 }, m_py{ 0 }, m_sx{ 0 }, m_sy{ 0 },
 		m_sw{ 0 }, m_sh{ 0 };
 
 	Link m_link;
-	f32 m_time{ 0.0f };
+	float m_time{ 0.0f };
 };
 
 #endif // SYG_NODE_CANVAS_H
